mypy, pylint, and pytest updates
