+++ pylitterbot/__init__.py
@@ -2,6 +2,9 @@
 __version__ = "2022.7.1b"
 
 from .account import Account
-from .robot import LitterRobot, LitterRobot3, LitterRobot4, Robot
+from .robot import Robot
+from .robot.litterrobot import LitterRobot
+from .robot.litterrobot3 import LitterRobot3
+from .robot.litterrobot4 import LitterRobot4
 
 __all__ = ["Account", "Robot", "LitterRobot", "LitterRobot3", "LitterRobot4"]
