+++ tests/conftest.py
@@ -8,7 +8,7 @@ import jwt
 import pytest
 from aioresponses import aioresponses
 
-from pylitterbot.robot import LR4_ENDPOINT
+from pylitterbot.robot.litterrobot4 import LR4_ENDPOINT
 from pylitterbot.session import LitterRobotSession
 
 from .common import (
